refactor(cli): separate process diagnostics from command execution
