+++ apps/cli/test/type/cliTypes.typecheck.ts
@@ -1,4 +1,5 @@
 import type { CliRunResult } from "../../src/cliTypes.js";
+import type { CommandDispatchReceiptResult } from "../../src/commands/command.js";
 
 const correlatedResult: CliRunResult = {
   code: 0,
@@ -24,7 +25,24 @@ const explicitUndefinedTrace: CliRunResult = {
   correlation: { status: "accepted", commandId: "cmd_type", traceId: undefined },
 };
 
+const acceptedReceiptResult: CommandDispatchReceiptResult = {
+  status: "accepted",
+  receipt: {
+    commandId: "cmd_type",
+    accepted: true,
+    status: "accepted",
+  },
+};
+
+// @ts-expect-error dispatch status and receipt acceptance are one discriminated fact.
+const mismatchedReceiptResult: CommandDispatchReceiptResult = {
+  status: "rejected",
+  receipt: acceptedReceiptResult.receipt,
+};
+
 void correlatedResult;
 void absentCorrelation;
 void explicitUndefinedCorrelation;
 void explicitUndefinedTrace;
+void acceptedReceiptResult;
+void mismatchedReceiptResult;
